MapFooter: fix legend icon showing incorrectly + refactor
`+hide shadow when collapsed`

`+fix bug - empty space when climbing layer off`

`+refactor`
