+++ src/components/Map/Map.tsx
@@ -10,9 +10,6 @@ import { SHOW_PROTOTYPE_UI } from '../../config';
 import { LayerSwitcherButton } from '../LayerSwitcher/LayerSwitcherButton';
 import { MaptilerLogo } from './MapFooter/MaptilerLogo';
 import { TopMenu } from './TopMenu/TopMenu';
-import { ClimbingLegend } from './MapFooter/ClimbingLegend';
-import { convertHexToRgba } from '../utils/colorUtils';
-import { usePersistedState } from '../utils/usePersistedState';
 
 const BrowserMap = dynamic(() => import('./BrowserMap'), {
   ssr: false,
@@ -56,18 +53,6 @@ const BottomRight = styled.div`
   z-index: 999;
 `;
 
-const FooterContainer = styled.div`
-  margin: 0 4px 4px 4px;
-  pointer-events: all;
-  border-radius: 8px;
-  padding: 6px;
-  box-shadow: 0 0 30px rgba(0, 0, 0, 0.3);
-  color: ${({ theme }) => theme.palette.text.primary};
-  background-color: ${({ theme }) =>
-    convertHexToRgba(theme.palette.background.paper, 0.5)};
-  backdrop-filter: blur(10px);
-`;
-
 const BugReportButton = () => (
   <Button size="small">
     <BugReport width="10" height="10" />
@@ -86,10 +71,7 @@ const NoscriptMessage = () => (
 const Map = () => {
   const [mapLoaded, setLoaded, setNotLoaded] = useBoolState(true);
   useEffect(setNotLoaded, []);
-  const [isLegendVisible, setIsLegendVisible] = usePersistedState<boolean>(
-    'isLegendVisible',
-    true,
-  );
+
   return (
     <>
       <BrowserMap onMapLoaded={setLoaded} />
@@ -102,16 +84,7 @@ const Map = () => {
       <BottomRight>
         {SHOW_PROTOTYPE_UI && <BugReportButton />}
         <MaptilerLogo />
-        <FooterContainer>
-          <ClimbingLegend
-            isLegendVisible={isLegendVisible}
-            setIsLegendVisible={setIsLegendVisible}
-          />
-          <MapFooter
-            isLegendVisible={isLegendVisible}
-            setIsLegendVisible={setIsLegendVisible}
-          />
-        </FooterContainer>
+        <MapFooter />
       </BottomRight>
     </>
   );
